WIP: add shim for request to support Streamlining.
The issue is basically that request "returns" (calls your callback with)
multiple non-error values -- response and body -- while Streamline expects
just one return value -- since it's making the code synchronous-like.

So, this shim wraps request's callbacks to "return" only one non-error value.
Instead of returning an array (tuple) or object (dictionary), I went the
route of adding the body as a property to the response object and returning
only the response.

The voodoo magic in this code is to preserve automatic "futures" support
that Streamline provides. That lets us parallelize requests if we want.