Update to latest node-request; remove patch.
@gasi, make sure to run `npm update request` or else code will fail now.

TODO: update other projects' package.jsons to require newer node-request too;
I'd noticed a spot or two which assumed it was monkey-patched for Streamline.

TODO also: we can now do just request('http://...') instead of request({uri}).

Unrelated to this, but while we're on the subject of TODOs: now that we're on
Neo4j 1.4, we should update our params to e.g. 'max_depth', not 'max depth'.