Build: pre-install compilers for robustness.
We need to be robust to this module getting installed by a parent that already
uses Streamline but not CoffeeScript. In that case, if we use the parent's
Streamline binary, it won't have CoffeeScript in its require() paths.

We work around that for now by forcing the local installation of CoffeeScript
and Streamline so that we know for certain the two will be siblings.

XXX This requires us to duplicate our semver spec for these dependencies.
TODO FIXME Revert this if npm changes to always install deps locally:

https://github.com/isaacs/npm/issues/2443