Remove profiling script; not needed anymore.
It was created when we were making the change from getters/setters on every
created Node/etc. instance to getters/setters on the prototypes only.

Now that that's stabilized, this simple profile isn't giving us any new or
useful information.

It'd still be nice to have some profiling at some point, but this is not it.